redesign-and-logic Added documentation, removed unnecessary imports, moved classes into their own files, created a class for commonly used colours, and fixed some formatting issues.
