+++ src/kivy_version/colour_converter_variables.py
@@ -1,0 +1,5 @@
+class ColourConverterVariables():
+    background_colour = "#000000"
+    default_output_colour = "#FFFFFF"
+    danger_colour = "#FF0000"
+    error_colour = "#FF726F"
