+++ src/kivy_version/window_manager.py
@@ -1,0 +1,8 @@
+from kivy.uix.screenmanager import ScreenManager
+
+class WindowManager(ScreenManager):
+    """
+        This is the class that handles the different interfaces that the 
+        application has, namely the RGBToHex and HexToRGB views.
+    """
+    pass
